+++ src/ChatCharacter.php
@@ -7,23 +7,22 @@ namespace HeroesofAbenez\Chat;
  * ChatCharacter
  *
  * @author Jakub Konečný
- * @property int|string $id
- * @property string $name
  */
 class ChatCharacter {
   use \Nette\SmartObject;
   
   /** @var int|string */
-  protected $id;
+  public $id;
   
   /**
    * @param int|string $id
    */
-  public function __construct($id, protected string $name) {
+  public function __construct($id, public string $name) {
     $this->id = $id;
   }
   
   /**
+   * @deprecated Access the property directly
    * @return int|string
    */
   public function getId() {
@@ -31,16 +30,23 @@ class ChatCharacter {
   }
 
   /**
+   * @deprecated Access the property directly
    * @param int|string $id
    */
   protected function setId($id): void {
     $this->id = $id;
   }
 
+  /**
+   * @deprecated Access the property directly
+   */
   public function getName(): string {
     return $this->name;
   }
 
+  /**
+   * @deprecated Access the property directly
+   */
   protected function setName(string $name): void {
     $this->name = $name;
   }